backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.15.0
djangorestframework-gis
djangorestframework-simplejwt
sqlparse==0.5.5
psycopg2
dj-database-url
dj-database-url==2.1.0
python-dotenv
qrcode[pil]
Pillow
django-cloudinary-storage
cloudinary